Delay showing splash screen until sizing complete #154
If the size of the window was different when the application was last
closed, the WindowManipulationService will place a resize event on the
dispatcher queue with the old size. The keyboard service will then resize
the window to the correct dimensions based on the user's startup
preferences. The way the dispatcher queue works means that messages are
not handled until a handler is attached and so the splash screen creation
is triggered too early when the window is still the old size.

To fix this, in the WindowManipulationService, we suppress the resize
message if we are initialising.

However, if the window is the correct size at startup then the call
from the keyboard service into the WindowManipulationService would assume
there was no work to do and wouldn't perform a resize.

So, we also remove checks to see whether we are already in the correct
state when resizing. This ensures the message from the keyboard startup
workflow is always raised and it is this message that triggers the splash
screen, when we are sized correctly.