Revert previous commit 'Remove repitition of word parameter when creating dictionary entry'
A single DictionaryEntry instance per word must be shared between the DictionaryService and the IAutoComplete implementation so that the usage counts are updated on the latter.

Storing this data in multiple places is not ideal, but a bigger issue than adding ngram support, so fixing it is deferred until later.