Support phrases with NGram Autocomplete #237
If there are spaces in the dictionary entry, then additionally treat the
combined initial characters as the word to be ngram-hashed. For example,
with "how are you", as well as the ngrams for the entire phrase, also
compute the ngrams for "hay" (the initial characters of each word) and add
the phrase to the dictionary for each one.