Enhances behaviour of WindowManipulationService when changing from floating to docking
OptiKey will now switch from floating to docked if:
It is expanded so that it touches 3 edges.
It is jumped to an edge it is already against (it will not dock if the
jump is diagonal as there is no good way to determine which edge to dock
against).

It will no longer change from floating to docked if:
It is moved diagonally into a corner of the screen as there is no good
way to determine which edge to dock against.