Update .gitignore
I'd rather the file was deleted in the solution rather than using git to ignore it. The problem is if the ignore is present, but the file is generated by Visual Studio you'll end up checking in a project file referencing app.config, which will not be committed, so the build will fail on the server.