Variable MaryTTS rate and volume
The rate and volume parameters are now sent to MaryTTS although there is
currently a bug in MaryTTS which prevents the rate from being changed. The
accepted volume range is 0.0 to 10.0 but higher than 1.4 gets disorted so
this has been set as the maximum (100%). The acceped rate range is 0.1 to
3.0 which it is converted to from the -10 to 10 rate option (but is
untested due to the aforementioned bug).