Correct Speak key behaviour with MaryTTS
When used with a MaryTTS voice, the Speak key now beahaves as it is
supposed to. This has been acomplished by querying how long the speech
will take then using async/await to trigger the onComplete action after
that time has elapsed (250 ms is added to the time to ensure it's
finished).

The list of MaryTTS voices has also been fixed. The list is delimited with
\r\n but previously only the \n was used to split the list so the \r was
left at the end of the list entries.