Add a user defined delay before speech
It may be benificial to have a small delay between selection of the Speak
key and when the speaking starts. This could be either because the key
selection sound is detracting from the speech of a single short word like
"no" or becuase a bluetooth speaker is in use and cuts off the beginning
of the speech. The delay has been limited to a maximum of five seconds.