CK always starts on top page and updated OBZ
The CK keyboard context is now set to top page on startup and access via
the main menu.

Updated the CK pageset.obz archive with yes/no on the top page, links to
the other keyboards on the leisure page, c/w/sh-ould on the question page
and 'I have pain in my...' acting as both a page key and a insert text
key.