Unpack CommuniKate pageset into the the AppData Roaming directory
The pageset cannot be unpacked into the Program Files (x86) directory
because OptiKey does not run as an administrator. When the pageset
location is set it triggers a restart of OptiKey and stages the
previously unpacked pageset to be deleted when OptiKey next starts.

The back icon used by CommuniKate has been moved from within the default
pageset archive to ./Resources/CommuniKate/ so that it can be used with
other pagesets.