Fix styling of empty CK keys.
Normally keys which aren't enabled (e.g. because they don't have a keyvalue
or are currently not relevant) are styled "greyed out", but this interferes with
the Communikate styling, which wants coloured blank keys.

On the original Communikate PR, this wasn't an issue since keys with null key
values coincidentally counted as "enabled", but for dynamic keyboards we
want the option of having no key value for an "empty" key.