Added more dps and have almost fully recovered the CK formatting
When a CK page is put to sleep the keys turn a solid colour instead of
black. This is because there is no longer a distinction between a blank
key and a sleeping key.

dps added:
- ForegroundColourOverride
- DisabledForegroundColourOverride
- HoverForegroundColourOverride
- SelectionProgressColourOverride
- SelectionForegroundColourOverride
- ShowOverScratchpadOveride

There is currently a bug when the conversation CK top page is first
navigated to. If it is put to sleep all the CK menu keys turn black
instead of their respective colours.