Use OverrideHasValue converter and transparent sleep indication
The Has<specific>ColourOverride properties have been removed in favour of
using a converter (OverrideHasValue) to check if the value is not null or
empty.

The disabled CK keys now have a transparent black layer over the top to
indicate they are disabled. This currently also effects all other keys
when sleeping but the reason has not yet been found.