Fix for an error in MultiKeySelection
As discussed in #401, when MultiKeySelection is enabled in a
language where the KeyValues.MultiKeySelectionKeys does not have a set
specified it crashes.

KeyValues.MultiKeySelectionKeys will now return an empty collection if the key
language is not found and the EnabledState of the MultiKeySelectionIsOnKey
is checked before multi-key selection is performed.