Added OBF file support and other enhancements
Now either an OBF file or an OBZ archive can be imported.

The button layout is now determined from the grid information. The
image locations are likewise determined by the defined image information.
For grids with five rows defined the first row is ignored because it
coresponds to the scratchpad row.