Save the path of the root board and some other fixes
The root board is now saved as Settings.Default.CommuniKateDefaultBoard so
that it can be directly referenced instead of setting the current board to
null.

The formatting for the first and second buttons were slightly muddled.
This caused the second button to be disabled when the first was and the
first button to not correctly fill with text when it didn't have an image.

Action buttons only have the black corner overlay if they navigate to
another board.

Updated the bundled pageset to reflect the changes made to how they are
supported.