Added the option to speak the text on CommuniKate buttons when selected
When this option is enabled, if the button doesn't have a sound, the
button's text will be spoken. This speach has its own volume and rate
settings so that it can be said quicker and quieter than normal speech.

Some other bits of code were also tidied-up (isMenuKey bits removed, etc).