SVG conversion, ":deleteword" support and handling paths with "../"
SVG images can now be viewed by first converting them to a PNG using the
Inkspace command-line conversion.

The OBF action ":deleteword" has been added to the list of recognised
actions and will delete the last word using FunctionKeys.BackMany.

Some pagesets contain paths with "../" in them. When the pagesets are
unpacked the paths are now stripped of any occurances of "../". The image,
sound and board paths from the OBF files are also stripped of any
occurances of "../" so that they match-up to the new file structure.