Media and other actions, enhanced back button and updated pageset
Added supported actions for media, volume and keyboard buttons:
	- ":ext_media_pause"
	- ":ext_media_next"
	- ":ext_media_previous"
	- ":ext_volume_up"
	- ":ext_volume_down"
	- ":ext_volume_mute"
	- ":ext_letters"
	- ":ext_numbers"
	- ":ext_mouse"

Buttons that contain actions are now identified by a marker in the top
left corner (similar to how buttons that lead to a different board are
idintified with a mark in the top right).

The included pageset has been updated to reflect these additions.

The Back button now will return to the previous keyboard if it is on the
root board.