Enhanced Back button and added support to Alphabetical keyboard
The Back button can now go back to any keyboard.

The CommuniKate board has been enabled for use with the Alphabetical
keyboard. An icon has been added for the CommuniKate key (it's the image
used to indicate the toppage).

Tidied the use of SvgViewboxs and Images on CK20Page.xaml to eliliminate
some errors.