Use ConversationAlpha1 for ComuniKate in fullscreen
The Communicate key on the ConversationAlpha1 keyboard led to the Alpha1
keboard. This meant that the scratchpad was mot shared properly.

The CK key on the ConversationNumericAndSymbols keyboard is now on the
same side as on the other keyboards and it now uses the same symbol.

The KeyValue CommuniKate has been renamed to CommuniKateKey to be in-line
with the other KeyValue names.