Make the text colour dependent upon the background colour
The sum of the background colour channels is calculated and the text
colour is set according to whether the sum is over or under half the
maximum (ie. white which is 3 * 0xFF).

This also takes into consideration the alpha channel and the applied
theme. When the alpha channel is less than max it will combine the
background theme colour to the colour channel sum in the proportion
defined by the alpha channel.