Show button borders and dynamicly set animation colours
The OBF button border definitions are now used to set a border around the
keys. The corner tags that indicate special keys now use the border
colour insead of black. This boarder is slightly thicker than the hover
border so that the two can be seen simultaneously.

The hover border colour and the key selection progress animation now use
either black or white depending on the border and background colours.