Fix CommuniKate settings and add Presage database settings
I had previously forgot to wire in the extra CK settings in the management
console.

The Presage database and number of suggestions can now be set from the
management console instead of having to edit the Presage config gile
manually. The ACAT presage database which includes apostrophe contractions
is now included and set as the default (it's zipped bacause it was too
large).