Add size/pos and calibration keys to mouse keyboard
This is a temporary measure to get round the issue that the mouse app
doesn't give you any access to the size/position settings, or the calibration.
Ultimately we'll probably move size/pos to settings, but we'll need to think
about how best to present these options to the user.

Also we have lots of planned changes to the mouse keyboard so watch this
space.