Add mechanism to hardcode installer strings into DLL
.NET's usual internationalisation architecture requires a satellite assembly, to keep
resources for different cultures separate from the compiled executable.
   This makes it difficult to expose translated strings to AdvancedInstaller, since we
can only attach a standalone DLL, and not all the associated resource folders.
   So as a workaround (the least bad option I could think of) we autogenerate a class
that contains all the translations for the (few) strings that need exposing to the
installer. This lets us keep all the translations inside the usual Optikey volunteer-
translations workflow.