Disable Aero snapping when resizing
Default behaviour in Windows means that resizing to the edge of
the screen results in snapping / maximising in that dimension.

The usual workaround is described here (and implemented in MainWindow
for drag re-positioning):
https://stackoverflow.com/a/13403067/9339328

But this doesn't work when resizing with the grabhandles with WindowChrome,
where the mouse events get eaten by the resize bar outside the client area.

This similar workaround is based on this instead:
https://stackoverflow.com/a/6657804/9339328