Make floating size-coercion more forgiving
Previously if Size & Position put any of Optikey outside the screen bounds,
everything would reset to the default (middle, 2/3 high, ...)

Now there's an attempt to cut off (insersect) at the screen bounds. If the
resulting rectangle still fails (either empty or "too small" then we fallback
to the default. This makes it easier for users to play around with sizes.