Make cursor flipping 'sticky'
Previously if you tried to select a point near the screen edge (right or bottom) then
the cursor could flip back and forth while you were trying to select, which was very
distracting. Now it stays flipped unless it needs to flip back on opposite side.

If people don't like the "non-standard" orientations then we could make it flip back
to "standard" over a larger region, so long as it's far enough away from the bottom
right threshold that it doesn't flip while trying to fixate.