Fix event that should have been filtered out
This was causing a double event to occur when using a gamepad
as trigger. When in SinglePoint selection mode, the "key" subscription
happened to receive a gamepad trigger event before the
"point" subscription, and although this was correctly filtered out in the
selection handler it was still bubbling through to the selectionresult
handler. Because there is an async delay between the two, the "point"
subscription had already acted on the point selection, and changed the
selection mode back to key selection, at which point the "key" selectionresult
handler thought it appropriate to do a "repeat last action". The user-facing
result was a double click the first time you tried a single click.

This change ensures the "key" trigger gets filtered out sooner, while the
selection mode is still "single point".

I have verified that if the events occur in the opposite order we observe
the correct behaviour. This is the case when a keyboard key is used as a
trigger.