Add char categories to fix Presage issues
Presage was:
- not working for contractions like "I'm"
- behaving a bit weirdly for compound words like "non-alphabetic".
- deleting adjacent symbols like parentheses when taking a suggestion

The CharCategories are used for 2 things:
- to cluster groups of chars for "BackMany" and
- to compute the in-progress word for word prediction

With this change we:
- Correctly treat "I'm" as a full word for prediction and deletion
- Treat "non-alphabetic" as 2 parts, with prediction after the hyphen
predicting the second part only
- As a consequence BackMany will be more conservative in situations
like "non-alphabetic" - this feels like a safe behavioural change
- Also, word prediction allows splits (with separate prediction) for
characters like "_" for programming variable names, as a bonus
- Note that for BackMany we treat a word separator like '-' in the same
way we treated a trailing space, e.g. 'non-alphabetic' will reduce to
'non-' and then BackMany will remove the hyphen and the 'non' together.
This is consistent with ctrl+backspace in many editors.

Note that we still treat symbols separately, so e.g. BackMany will take
4 operations to delete `(aside) ` - one space, one ), one word, one (.